+++ examples/autoconfig_mx.rs
@@ -48,12 +48,10 @@ fn main() {
     };
 
     for record in records {
-        // TODO: restore when the domain new API is released:
-        // println!("{} {}", record.rdata.preference.get(), record.rdata.exchange);
         println!(
             "{} {}",
-            record.data().preference(),
-            record.data().exchange()
+            record.rdata.preference.get(),
+            record.rdata.exchange
         );
     }
 }